# swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"HouseResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"imageUrl": {
						"type": "string"
					},
					"caption": {
						"type": "string",
						"nullable": true
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"size": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"hasLiked": {
						"type": "boolean"
					},
					"hasOrdered": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"imageUrl",
					"caption",
					"price",
					"address",
					"state",
					"city",
					"zipCode",
					"size",
					"createdAt",
					"userId",
					"username",
					"avatarUrl",
					"hasLiked",
					"hasOrdered"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateHouseBase64Input": {
				"properties": {
					"imageBase64": {
						"type": "string"
					},
					"imageFileType": {
						"type": "string"
					},
					"caption": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"zipCode": {
						"type": "string"
					},
					"size": {
						"type": "string"
					}
				},
				"required": [
					"imageBase64",
					"imageFileType",
					"price",
					"address",
					"state",
					"city",
					"zipCode",
					"size"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"OrderResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"houseId": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"houseImageUrl": {
						"type": "string"
					},
					"houseCaption": {
						"type": "string"
					},
					"houseCity": {
						"type": "string"
					},
					"houseState": {
						"type": "string"
					},
					"houseAddress": {
						"type": "string"
					},
					"houseZipCode": {
						"type": "string"
					},
					"houseSize": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"userId",
					"houseId",
					"username",
					"avatarUrl",
					"createdAt",
					"houseImageUrl",
					"houseCaption",
					"houseCity",
					"houseState",
					"houseAddress",
					"houseZipCode",
					"houseSize"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateOrderInput": {
				"properties": {
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"startDate",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Airbnb Simplified API",
		"version": "1.0.0",
		"description": "API docs for simplified Airbnb backend",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Your Name"
		}
	},
	"paths": {
		"/houses": {
			"post": {
				"operationId": "CreateHouse",
				"responses": {
					"200": {
						"description": "Post Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HouseResponse"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Houses"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateHouseBase64Input"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetFeedHouses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/HouseResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Houses"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/houses/search": {
			"get": {
				"operationId": "SearchHouses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/HouseResponse"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Houses"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"default": "",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "city",
						"required": false,
						"schema": {
							"default": "",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "zipCode",
						"required": false,
						"schema": {
							"default": "",
							"type": "string"
						}
					}
				]
			}
		},
		"/houses/{houseId}": {
			"get": {
				"operationId": "GetHouseById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HouseResponse"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Houses"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "houseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/houses/{houseId}/like": {
			"post": {
				"operationId": "LikeHouse",
				"responses": {
					"201": {
						"description": "Liked",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Interactions (Likes & Orders)"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "houseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/houses/{houseId}/unlike": {
			"delete": {
				"operationId": "UnlikeHouse",
				"responses": {
					"200": {
						"description": "Unliked",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Interactions (Likes & Orders)"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "houseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/houses/{houseId}/orders": {
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"201": {
						"description": "Order Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderResponse"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Interactions (Likes & Orders)"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "houseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOrderInput"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OrderResponse"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Interactions (Likes & Orders)"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "houseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}